backend/pyproject.toml
[project]
name = "forecastle-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "joblib>=1.5.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "prophet>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
